fullstack
fullstack

+++ src/main/java/com/csa/samplefullstack/entity/Employee.java
@@ -1,4 +1,4 @@
-package com.csa.samplefullstack.entity;
+ package com.csa.samplefullstack.entity;
 
 import javax.persistence.*;
 
@@ -8,12 +8,12 @@ public class Employee {
  
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int employeeId;
+    private Long employeeId;
     private String name;
     private String department;
     private String location;
 
-    public Employee(int employeeId, String name, String department, String location) {
+    public Employee(Long employeeId, String name, String department, String location) {
         this.employeeId = employeeId;
         this.name = name;
         this.department = department;
@@ -24,11 +24,11 @@ public class Employee {
         
     }
     
-    public int getEmployeeId() {
+    public Long getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(int employeeId) {
+    public void setEmployeeId(Long employeeId) {
         this.employeeId = employeeId;
     }
